displayed textinput and buttons;
