+++ android/app/build.gradle
@@ -137,7 +137,7 @@ android {
 }
 
 dependencies {
-    compile project(':react-native-cookies')
+    compile project(':react-native-maps')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
@@ -149,7 +149,7 @@ dependencies {
         exclude group: 'com.google.android.gms', module: 'play-services-maps'
     }
     implementation 'com.google.android.gms:play-services-base:10.0.1'
-    implementation 'com.google.android.gms:play-services-maps:10.0.1'
+    implementation 'com.google.android.gms:play-services-maps:10.2.4'
 }
 
 // Run this once to be able to run the application with BUCK
