+++ android/app/src/main/java/com/newworld/MainApplication.java
@@ -3,7 +3,8 @@ package com.newworld;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
-import com.psykar.cookiemanager.CookieManagerPackage;
+import com.airbnb.android.react.maps.MapsPackage;
+// import com.psykar.cookiemanager.CookieManagerPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
@@ -25,7 +26,6 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
-          new CookieManagerPackage()
           new MapsPackage()
       );
     }
