+++ android/settings.gradle
@@ -1,6 +1,6 @@
 rootProject.name = 'NewWorld'
-include ':react-native-cookies'
-project(':react-native-cookies').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-cookies/android')
+include ':react-native-maps'
+project(':react-native-maps').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-maps/lib/android')
 
 include ':app'
 include ':react-native-maps'
