+++ src/App.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react";
 
 import Login from "./components/Login";
-
+import Map from "./components/Map";
 export default () => (
-  <Login />
+  // <Login />
+  <Map />
 );
